chore: remove deprecated `Signature` (#899)
* chore: remove deprecated Signature

* fix doc
