+++ crates/primitives/src/lib.rs
@@ -54,9 +54,9 @@ mod signed;
 pub use signed::{BigIntConversionError, ParseSignedError, Sign, Signed};
 
 mod signature;
-pub use signature::{normalize_v, to_eip155_v, PrimitiveSignature, SignatureError};
 #[allow(deprecated)]
-pub use signature::{Parity, Signature};
+pub use signature::PrimitiveSignature;
+pub use signature::{normalize_v, to_eip155_v, Signature, SignatureError};
 
 pub mod utils;
 pub use utils::{eip191_hash_message, keccak256, Keccak256};
